move polynomial regression to lib
